+++ packages/analytics/analytics-metric-provider/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.1.1](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-metric-provider@4.1.0...@kong-ui-public/analytics-metric-provider@4.1.1) (2024-03-08)
+
+
+### Bug Fixes
+
+* replace CopyUuid with KCopy [khcp-9465] ([#1210](https://github.com/Kong/public-ui-components/issues/1210)) ([e409d60](https://github.com/Kong/public-ui-components/commit/e409d60b7ab483c200ed6bb33f43ff571d6e6b88))
+
+
+
+
+
+# [4.1.0](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-metric-provider@4.0.11...@kong-ui-public/analytics-metric-provider@4.1.0) (2024-03-07)
+
+
+### Features
+
+* **analytics:** get config from analytics endpoint [MA-2532] ([#1208](https://github.com/Kong/public-ui-components/issues/1208)) ([3a1ae09](https://github.com/Kong/public-ui-components/commit/3a1ae09827dd5797d4fd5e93fe60935758c1c6bc))
+
+
+
+
+
 ## [4.0.11](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-metric-provider@4.0.10...@kong-ui-public/analytics-metric-provider@4.0.11) (2024-03-07)
 
 **Note:** Version bump only for package @kong-ui-public/analytics-metric-provider